+++ paver/tests/other_pavement.py
@@ -1,0 +1,9 @@
+from paver.easy import *
+
+from paver.tests import test_tasks
+
+options(foo=1)
+
+@task
+def t1(options):
+    test_tasks.OP_T1_CALLED = options.foo